requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
customtkinter>=5.2.0
pillow>=10.0.0  # Updated for security
send2trash>=1.8.0
tqdm>=4.65.0
pywin32==310   # Pinned to avoid conflicts
pyinstaller==6.13.0  # Pinned for stability
tzdata          # Optional (only needed on Windows)